+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [0ver](https://0ver.org) (more or less).
 
 ## [Unreleased]
 
+### Changed
+
+- Renamed flag `--github-org` to `--github-orgs`
+- Renamed flag `--gitlab-group` to `--gitlab-groups`
+
 ## [v0.0.1] - 2021-03-22
 
 ### Added